Move route to http endpoint so it can be run without auth
